+++ spec/controllers/nippos/backs_controller_spec.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+RSpec.describe Nippos::BacksController do
+  describe 'POST create' do
+    it 'shows form' do
+      post :create, back: '戻る', nippo: {
+        reported_for: Time.zone.today,
+        body: FFaker::Lorem.paragraph,
+      }
+      expect(response).to have_http_status(:success)
+      expect(response).to render_template(:new)
+    end
+  end
+
+  describe 'PATCH update' do
+    let(:nippo) { FG.create(:nippo) }
+
+    it 'shows form' do
+      patch :update, id: nippo.id, back: '戻る', nippo: { body: 'changed' }
+      expect(response).to have_http_status(:success)
+      expect(response).to render_template(:show)
+    end
+
+    context 'nippo has already sent' do
+      before { nippo.update(status: :sent) }
+
+      it 'redirects show' do
+        patch :update, id: nippo.id, back: '戻る', nippo: { body: 'changed' }
+        expect(response).to redirect_to(nippo_path(nippo))
+      end
+    end
+  end
+end
