--- spec/support/simplecov.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-require 'simplecov'
-# SimpleCov.minimum_coverage 90
-SimpleCov.start 'rails'
